[REFACTOR] 안쓰는클래스 및 경로 수정
